Utility + hook to capture max value calculations
